Updated german translation
At first added all missing translations, secondly fixed grammar and
spelling mistakes and third switched from mixed formal and informal
style to a consequent informal style (in German there is the formal
"Sie" and the informal "Du", both mean you in English) as this fits
better to the context of a game. Further improved the language stlye to
be more fluent in some points.